Avoid fast-math errors when reading hypergammon databases
Decode 24-bit hypergammon database values using double division
before converting them to float.

configure.ac enables -ffast-math by default for GCC builds.  With this
option, GCC may replace the float division with reciprocal
multiplication, changing decoded values and causing the range assertion
in StartFromDatabase() to fail even when the resulting float is 1.0.

Use the same decoding in both StartFromDatabase() and
ReadHypergammon().

+++ bearoff.c
@@ -348,13 +348,13 @@ ReadHypergammon(const bearoffcontext * pbc, const unsigned int iPos, float arOut
     if (arOutput)
         for (i = 0; i < NUM_OUTPUTS; ++i) {
             us = pc[3 * i] | (pc[3 * i + 1]) << 8 | (pc[3 * i + 2]) << 16;
-            arOutput[i] = (float) us / 16777215.0f;
+            arOutput[i] = (float) ((double) us / 16777215.0);
         }
 
     if (arEquity)
         for (i = 0; i < 4; ++i) {
             us = pc[15 + 3 * i] | (pc[15 + 3 * i + 1]) << 8 | (pc[15 + 3 * i + 2]) << 16;
-            arEquity[i] = ((float) us / 16777215.0f - 0.5f) * 6.0f;
+            arEquity[i] = ((float) ((double) us / 16777215.0) - 0.5f) * 6.0f;
         }
 
     return 0;
